save state (not runnable)

+++ src/main/java/de/micromata/jira/rest/core/domain/CommentsBean.java
@@ -1,21 +1,20 @@
 package de.micromata.jira.rest.core.domain;
 
-import com.google.gson.annotations.Expose;
-
 import java.util.ArrayList;
 import java.util.List;
 
+import com.google.gson.annotations.Expose;
+
 
 public class CommentsBean {
-
     @Expose
     private List<CommentBean> comments = new ArrayList<>();
     @Expose
-    private Integer maxResults;
+    private long              maxResults;
     @Expose
-    private Integer startAt;
+    private long              startAt;
     @Expose
-    private Integer total;
+    private long              total;
 
     public List<CommentBean> getComments() {
         return comments;
@@ -25,27 +24,27 @@ public class CommentsBean {
         this.comments = comments;
     }
 
-    public Integer getMaxResults() {
+    public long getMaxResults() {
         return maxResults;
     }
 
-    public void setMaxResults(Integer maxResults) {
+    public void setMaxResults(long maxResults) {
         this.maxResults = maxResults;
     }
 
-    public Integer getStartAt() {
+    public long getStartAt() {
         return startAt;
     }
 
-    public void setStartAt(Integer startAt) {
+    public void setStartAt(long startAt) {
         this.startAt = startAt;
     }
 
-    public Integer getTotal() {
+    public long getTotal() {
         return total;
     }
 
-    public void setTotal(Integer total) {
+    public void setTotal(long total) {
         this.total = total;
     }
 }
